signing out completed and helper method created
