+++ app/views/sessions/new.erb
@@ -2,10 +2,10 @@
 
 <form action='/sessions' method='POST'>
 	<label for='email'>
-		Email: <input name='email' type='email'>
+		Email: <input name='email' type='email' placeholder='email'>
 	</label>
 	<label for='password'>
-		Password: <input name='password' type='password'>
+		Password: <input name='password' type='password' placeholder='password'>
 	</label>
-	<input type='submit' value='Sign In'
+	<input type='submit' value='Sign In'>
 </form>
