+++ spec/features/user_management_spec.rb
@@ -1,15 +1,5 @@
 feature 'user sign up' do
 
-  def sign_up(email: 'rich@gmail.com',
-              password: '12345678',
-              password_confirmation: '12345678')
-    visit '/users/new'
-    fill_in :email, with: email
-    fill_in :password, with: password
-    fill_in :password_confirmation, with: password_confirmation
-    click_button 'Sign Up'
-  end
-
   scenario 'I can sign up as a new user' do
     expect { sign_up }.to change(User, :count).by 1
     expect(page).to have_content('Welcome, rich@gmail.com')
@@ -53,15 +43,24 @@ feature 'user can sign in' do
 			    password_confirmation: '12345678')
 	end
 
-	def sign_in(email:, password:)
-		visit '/sessions/new'
-		fill_in 'email', with: email
-		fill_in 'password', with: password
-		click_button 'Sign In'
-	end
-
 	scenario 'with the correct credentails' do
 		sign_in(email: user.email, password: user.password)
 		expect(page).to have_content "Welcome, #{user.email}"
+	end
+end
+
+feature 'users can log out' do
+
+	before(:each) do
+		User.create(email: 'rich@gmail.com',
+			    password: '12345678',
+			    password_confirmation: '12345678')
+	end
+
+	scenario 'while being signed in' do
+		sign_in(email: 'rich@gmail.com', password: '12345678')
+		click_button 'Sign Out'
+		expect(page).to have_content('Thanks for visiting, come back soon')
+		expect(page).not_to have_content('Welcome, rich@gmail.com')
 	end
 end
