+++ spec/helpers/session.rb
@@ -1,0 +1,20 @@
+module SessionHelpers
+	
+	def sign_in(email:, password:)
+		visit '/sessions/new'
+		fill_in 'email', with: email
+		fill_in 'password', with: password
+		click_button 'Sign In'
+	end
+
+	def sign_up(email: 'rich@gmail.com',
+		    password: '12345678',
+		    password_confirmation: '12345678')
+       		visit '/users/new'
+		fill_in :email, with: email
+		fill_in :password, with: password
+		fill_in :password_confirmation, with: password_confirmation
+		click_button 'Sign Up'
+  	end
+end
+
