+++ spec/spec_helper.rb
@@ -6,6 +6,7 @@ require './app/app'
 require "codeclimate-test-reporter"
 require 'database_cleaner'
 require 'web_helper'
+require_relative 'helpers/session'
 
 Capybara.app = Chitter_Challenge
 CodeClimate::TestReporter.start
@@ -14,6 +15,10 @@ RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
+
+RSpec.configure do |config|
+  config.include SessionHelpers
+end
 
   config.mock_with :rspec do |mocks|
     mocks.verify_partial_doubles = true
